Collocate state and handlers in home component
